[seller] refactor(228): header 영양 성분 정보 constant를 entity에서 features로 이동, 변수명 수정
