--- apps/seller/src/entity/products/create/create-header/category-options.constants.ts
@@ -1,16 +1,0 @@
-import { OptionTags } from './options-tag.type'
-
-export const CategoryOptions: OptionTags[] = [
-  {
-    title: '칼로리 다운',
-    tooltip: '저당, 저지방 성분 기준을 충족시 제품에 표시됩니다.',
-  },
-  {
-    title: '단백질 듬뿍',
-    tooltip: '고단백 성분 기준을 충족시 제품에 표시됩니다.',
-  },
-  {
-    title: '속 편한 즐거움',
-    tooltip: '비건과 글루텐프리 성분 기준을 충족시 제품에 표시됩니다.',
-  },
-]
