--- apps/seller/src/entity/products/create/create-header/essential-options.constants.ts
@@ -1,24 +1,0 @@
-import { OptionTags } from './options-tag.type'
-
-export const EssentialOptions: OptionTags[] = [
-  {
-    title: '글루텐프리',
-    tooltip: '밀·보리·호밀 등 글루텐 함유 곡물 미포함',
-  },
-  {
-    title: '비건',
-    tooltip: '동물성 성분 전혀 없음',
-  },
-  {
-    title: '고단백',
-    tooltip: '100g당 11g 이상',
-  },
-  {
-    title: '저지방',
-    tooltip: '100g당 3g 미만 / 100ml당 1.5g 미만',
-  },
-  {
-    title: '저당',
-    tooltip: '100g당 5g 미만 / 100ml당 2.5g 미만',
-  },
-]
