--- apps/seller/src/entity/products/create/create-header/options-tag.type.ts
@@ -1,4 +1,0 @@
-export type OptionTags = {
-  title: string
-  tooltip: string
-}
