--- apps/seller/src/entity/products/index.ts
@@ -1,4 +1,0 @@
-// Create - Header
-export * from './create/create-header/category-options.constants'
-export * from './create/create-header/essential-options.constants'
-export * from './create/create-header/options-tag.type'
