+++ apps/seller/src/features/products/create/create-header/index.ts
@@ -1,2 +1,4 @@
 export { ProductHeader } from './create-header.ui'
 export { ProductHeaderTags } from './create-header-tags.ui'
+export { ESSENTIAL_OPTIONS, CATEGORY_OPTIONS } from './create-header.constant'
+export type { OptionTags } from './create-header.constant'
